--- components/index.ts
@@ -1,1 +1,0 @@
-// dummy file to keep typescript language server happy
