+++ package.json
@@ -14,13 +14,16 @@
     "nuxt": "^3.0.0"
   },
   "scripts": {
-    "dev": "nuxt dev",
-    "build": "nuxt build",
-    "generate": "nuxt generate",
-    "start": "nuxt start"
+    "dev": "nuxi dev",
+    "build": "nuxi build",
+    "generate": "nuxi generate",
+    "start": "nuxi start",
+    "typecheck": "nuxi typecheck"
   },
   "devDependencies": {
     "sass": "^1.61.0",
-    "sass-loader": "^13.2.2"
+    "sass-loader": "^13.2.2",
+    "typescript": "^5.0.4",
+    "vue-tsc": "^1.6.3"
   }
 }
